// src/components/table/control-panel/filter/lib/types.ts
export interface IFilterItem {
  title: string;
  type: 'input' | 'radio';
  stateKey?: string;
  defaultValue?: string;
  radioItems?: RadioItem[];
  placeholder?: string;
}

type RadioItem = {
  value: string;
  label: string;
  className?: string;
};

export interface IFilter {
  login: string;
  language: string;
}

export interface IStateRecord {
  [key: string]: { value: string; onChange: (_: unknown) => void };
}